microservices/order-service/src/requirements.txt
fastapi
uvicorn[standard]
pika
pydantic
python-dotenv
